Feat: Refactor Structure of directory
